Move auto random seeding to hardware specific implementation files
(This is expected to increase flash size, marginally, but the effect will mostly go
away once implementation and usage are once again in the same compilation unit.)
