+++ MozziGuts_impl_ESP32.hpp
@@ -161,4 +161,12 @@ void stopMozzi() {
 }
 //// END AUDIO OUTPUT code ///////
 
+//// BEGIN Random seeding ////////
+void autoRandomSeeds(uint32_t *x, uint32_t *y, uint32_t *z) {
+  *x = esp_random();
+  *y = esp_random();
+  *z = esp_random();
+}
+//// END Random seeding ////////
+
 #undef ESP_SAMPLE_SIZE    // only used inside this file
