+++ MozziGuts_impl_MBED.hpp
@@ -221,5 +221,11 @@ void stopMozzi() {
 #endif
 ////// END audio output code //////
 
+//// BEGIN Random seeding ////////
+void autoRandomSeeds(uint32_t *x, uint32_t *y, uint32_t *z) {
+#warning Automatic random seedings is not implemented on this platform
+}
+//// END Random seeding ////////
+
 #undef CHUNKSIZE
 #undef US_PER_AUDIO_TICK
