+++ MozziGuts_impl_RENESAS.hpp
@@ -168,3 +168,9 @@ void stopMozzi() {
 #endif
 }
 //// END AUDIO OUTPUT code ///////
+
+//// BEGIN Random seeding ////////
+void autoRandomSeeds(uint32_t *x, uint32_t *y, uint32_t *z) {
+#warning Automatic random seedings is not implemented on this platform
+}
+//// END Random seeding ////////
