+++ MozziGuts_impl_RP2040.hpp
@@ -266,5 +266,11 @@ void stopMozzi() {
 }
 ////// END audio output code //////
 
+//// BEGIN Random seeding ////////
+void autoRandomSeeds(uint32_t *x, uint32_t *y, uint32_t *z) {
+#warning Automatic random seedings is not implemented on this platform
+}
+//// END Random seeding ////////
+
 #undef MOZZI_RP2040_BUFFERS
 #undef MOZZI_RP2040_BUFFER_SIZE
