+++ MozziGuts_impl_STM32duino.hpp
@@ -171,3 +171,9 @@ void stopMozzi() {
 }
 
 //// END AUDIO OUTPUT code ///////
+
+//// BEGIN Random seeding ////////
+void autoRandomSeeds(uint32_t *x, uint32_t *y, uint32_t *z) {
+#warning Automatic random seedings is not implemented on this platform
+}
+//// END Random seeding ////////
